requirements.txt
# AI Research Assistant - Requirements

# Core Dependencies
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Optional Dependencies (for enhanced features)
google-generativeai==0.3.1
python-dotenv==1.0.0

# Utility Dependencies
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4

# HTML Processing
soupsieve==2.5

# JSON handling (built-in but listed for reference)
# json - built-in
# datetime - built-in
# typing - built-in
